【WA】[第一回日本最強プログラマー学生選手権-予選-]B - Kleene Inversion

+++ jsc2019-qual/jsc2019_qual_b.cpp
@@ -155,7 +155,7 @@ int main() {
     }
 
     mint c = 0;
-    for (int i : min_count) {
+    for (ll i : min_count) {
         c += i * k * (k - 1) / 2;
     }
